stacks/backend/fastapi-postgresql/requirements.txt
# Production Dependencies for FastAPI PostgreSQL Backend
# Following gold standard practices for dependency management

# Core FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy[asyncio]>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client
httpx>=0.25.0

# Configuration
python-dotenv>=1.0.0

# Caching (optional)
redis>=5.0.0

# Monitoring
prometheus-client>=0.19.0
structlog>=23.2.0

# Utilities
python-dateutil>=2.8.0
aiofiles>=23.2.0

# Optional: Background Tasks
celery>=5.3.0
